[yeek] Refactor "test.lib.test_util" -> "api.test.test_util"
