+++ src/mod/elona/test/api/Calc.lua
@@ -1,7 +1,7 @@
 local Calc = require("mod.elona.api.Calc")
 local Rand = require("api.Rand")
 local Assert = require("api.test.Assert")
-local test_util = require("test.lib.test_util")
+local test_util = require("api.test.test_util")
 local Rank = require("mod.elona.api.Rank")
 local save = require("internal.global.save")
 local Area = require("api.Area")
