+++ src/mod/elona/test/event/feat.lua
@@ -1,5 +1,5 @@
 local Area = require("api.Area")
-local test_util = require("test.lib.test_util")
+local test_util = require("api.test.test_util")
 local Map = require("api.Map")
 local ElonaCommand = require("mod.elona.api.ElonaCommand")
 local Assert = require("api.test.Assert")
