+++ src/mod/elona_sys/test/api/Skill.lua
@@ -1,6 +1,6 @@
 local Skill = require("mod.elona_sys.api.Skill")
 local Assert = require("api.test.Assert")
-local test_util = require("test.lib.test_util")
+local test_util = require("api.test.test_util")
 
 function test_Skill_modify_resist_level()
    local chara = test_util.stripped_chara("elona.putit")
