+++ src/test/unit/api/Chara.lua
@@ -1,5 +1,5 @@
 local Chara = require("api.Chara")
-local test_util = require("test.lib.test_util")
+local test_util = require("api.test.test_util")
 local Map = require("api.Map")
 local Assert = require("api.test.Assert")
 local save = require("internal.global.save")
