+++ src/test/unit/api/IStackableObject.lua
@@ -2,7 +2,7 @@ local Item = require("api.Item")
 local Assert = require("api.test.Assert")
 local InstancedMap = require("api.InstancedMap")
 local Chara = require("api.Chara")
-local test_util = require("test.lib.test_util")
+local test_util = require("api.test.test_util")
 
 function test_IStackableObject_separate__single()
    local item = test_util.stripped_item("elona.putitoro")
