+++ src/test/unit/api/activity/IActivity.lua
@@ -4,7 +4,7 @@ local Activity = require("api.Activity")
 local Assert = require("api.test.Assert")
 local I18N = require("api.I18N")
 local Chara = require("api.Chara")
-local test_util = require("test.lib.test_util")
+local test_util = require("api.test.test_util")
 local Item = require("api.Item")
 local MapObject = require("api.MapObject")
