+++ src/test/unit/api/chara/ICharaSkills.lua
@@ -1,5 +1,5 @@
 local Assert = require("api.test.Assert")
-local test_util = require("test.lib.test_util")
+local test_util = require("api.test.test_util")
 
 function test_ICharaSkills_init__prototype_default_skills()
    local chara = test_util.stripped_chara("elona.zeome")
