+++ src/test/unit/event/instantiate_mef.lua
@@ -1,6 +1,6 @@
 local InstancedMap = require("api.InstancedMap")
 local Chara = require("api.Chara")
-local test_util = require("test.lib.test_util")
+local test_util = require("api.test.test_util")
 local Assert = require("api.test.Assert")
 local data = require("internal.data")
 local Mef = require("api.Mef")
