+++ src/test/unit/event/map_init.lua
@@ -1,5 +1,5 @@
 local Area = require("api.Area")
-local test_util = require("test.lib.test_util")
+local test_util = require("api.test.test_util")
 local Item = require("api.Item")
 local Assert = require("api.test.Assert")
 local Map = require("api.Map")
