+++ src/test/unit/internal/object.lua
@@ -1,7 +1,7 @@
 local Chara = require("api.Chara")
 local Assert = require("api.test.Assert")
 local InstancedMap = require("api.InstancedMap")
-local test_util = require("test.lib.test_util")
+local test_util = require("api.test.test_util")
 local Item = require("api.Item")
 local Inventory = require("api.Inventory")
 local Map = require("api.Map")
