+++ src/test/unit/location.lua
@@ -1,4 +1,4 @@
-local test_util = require("test.lib.test_util")
+local test_util = require("api.test.test_util")
 local InstancedMap = require("api.InstancedMap")
 local Chara = require("api.Chara")
 local Item = require("api.Item")
