+++ src/tools/cli/commands/test.lua
@@ -14,7 +14,7 @@ local config_store = require("internal.config_store")
 local save_store = require("internal.save_store")
 local SaveFs = require("api.SaveFs")
 local Stopwatch = require("api.Stopwatch")
-local test_util = require("test.lib.test_util")
+local test_util = require("api.test.test_util")
 local Advice = require("api.Advice")
 local main_state = require("internal.global.main_state")
 
@@ -44,7 +44,7 @@ local function reset_all_globals()
    save_store = require("internal.save_store")
    SaveFs = require("api.SaveFs")
    Stopwatch = require("api.Stopwatch")
-   test_util = require("test.lib.test_util")
+   test_util = require("api.test.test_util")
    Advice = require("api.Advice")
    main_state = require("internal.global.main_state")
 end
